add documentation to all the public functions
