+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransferSchedule.java
@@ -16,8 +16,13 @@ import java.nio.ByteBuffer;
 @Getter
 @EqualsAndHashCode(callSuper = true)
 public class TransferSchedule extends Payload {
-
+    /**
+     * The account address of the recepient.
+     */
     private final AccountAddress to;
+    /**
+     * The release schedule. This can be at most 255 elements.
+     */
     private final Schedule[] amount;
     private final UInt64 maxEnergyCost;
 
@@ -27,15 +32,28 @@ public class TransferSchedule extends Payload {
         this.maxEnergyCost = maxEnergyCost;
     }
 
+    static TransferSchedule createNew(AccountAddress to, Schedule[] amount, UInt64 maxEnergyCost) {
+        return new TransferSchedule(to, amount, maxEnergyCost);
+    }
+
+    /**
+     * This is a method that returns the type of the payload.
+     */
     @Override
     public PayloadType getType() {
         return PayloadType.TRANSFER_WITH_SCHEDULE;
     }
 
+    /**
+     * This function returns the transaction type of the transaction.
+     */
     public byte getTransactionType() {
         return TransactionType.TRANSFER_WITH_SCHEDULE.getValue();
     }
 
+    /**
+     * @return The byte array of the transaction.
+     */
     @Override
     byte[] getBytes() {
         val schedule_len = amount.length;
@@ -46,7 +64,7 @@ public class TransferSchedule extends Payload {
 
         buffer.put(to.getBytes());
 
-        buffer.put((byte)schedule_len);
+        buffer.put((byte) schedule_len);
         for (int i = 0; i < schedule_len; i++) {
             val schedule_buffer = amount[i].getBytes();
             buffer.put(schedule_buffer);
@@ -57,9 +75,5 @@ public class TransferSchedule extends Payload {
     @Override
     UInt64 getTransactionTypeCost() {
         return this.maxEnergyCost;
-    }
-
-    static TransferSchedule createNew(AccountAddress to, Schedule[] amount, UInt64 maxEnergyCost) {
-        return new TransferSchedule(to, amount, maxEnergyCost);
     }
 }
