+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransferScheduleWithMemo.java
@@ -15,11 +15,17 @@ import java.nio.ByteBuffer;
 @Getter
 @EqualsAndHashCode(callSuper = true)
 public final class TransferScheduleWithMemo extends Payload {
-
+    /**
+     * The account address of the recepient.
+     */
     private final AccountAddress to;
+    /**
+     * The release schedule. This can be at most 255 elements.
+     */
     private final Schedule[] amount;
-    private final UInt64 maxEnergyCost;
     private final Memo memo;
+
+    private final UInt64 maxEnergyCost;
 
     public TransferScheduleWithMemo(AccountAddress to, Schedule[] amount, Memo memo, UInt64 maxEnergyCost) {
         this.to = to;
@@ -28,15 +34,28 @@ public final class TransferScheduleWithMemo extends Payload {
         this.maxEnergyCost = maxEnergyCost;
     }
 
+    static TransferScheduleWithMemo createNew(AccountAddress to, Schedule[] amount, Memo memo, UInt64 maxEnergyCost) {
+        return new TransferScheduleWithMemo(to, amount, memo, maxEnergyCost);
+    }
+
+    /**
+     * This is a method that returns the type of the payload.
+     */
     @Override
     public PayloadType getType() {
         return PayloadType.TRANSFER_WITH_SCHEDULE_AND_MEMO;
     }
 
+    /**
+     * This function returns the transaction type of this transaction.
+     */
     public byte getTransactionType() {
         return TransactionType.TRANSFER_WITH_SCHEDULE_AND_MEMO.getValue();
     }
 
+    /**
+     * @return The byte array of the transaction.
+     */
     @Override
     byte[] getBytes() {
         val schedule_len = amount.length;
@@ -47,7 +66,7 @@ public final class TransferScheduleWithMemo extends Payload {
         buffer.put(to.getBytes());
         buffer.put(memo.getBytes());
 
-        buffer.put((byte)schedule_len);
+        buffer.put((byte) schedule_len);
         for (int i = 0; i < schedule_len; i++) {
             val schedule_buffer = amount[i].getBytes();
             buffer.put(schedule_buffer);
@@ -59,9 +78,5 @@ public final class TransferScheduleWithMemo extends Payload {
     @Override
     UInt64 getTransactionTypeCost() {
         return this.maxEnergyCost;
-    }
-
-    static TransferScheduleWithMemo createNew(AccountAddress to, Schedule[] amount, Memo memo, UInt64 maxEnergyCost) {
-        return new TransferScheduleWithMemo(to, amount, memo, maxEnergyCost);
     }
 }
